Fix the local type environment of module finalizers.
Summary:
Before, the local type environment of the finalizer was the one present
when the finalizer was added. This ignored the top-level Ids that would
be present in the local type environment at the end of type-checking.
This behavior was introduced in the fix to #11832.

Now the local type environment at the end of type-checking is used to
extend the local type environment of the finalizer.

Test Plan: ./validate

Reviewers: simonpj, bgamari, austin

Subscribers: mboes, thomie

Differential Revision: https://phabricator.haskell.org/D2519

GHC Trac Issues: #11832